requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.2.0
django-import-export==1.2.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.9
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.2
paramiko==2.6.0
Pillow==7.1.1
pycparser==2.19
Pygments==2.6.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.1.2
requests==2.23.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.1
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
